aiBackend/requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
jinja2>=3.1.0 